refactor: interface design
