+++ internal/sbi/processor/processor.go
@@ -5,11 +5,9 @@ import (
 	"github.com/free5gc/nssf/pkg/factory"
 )
 
-// TODO: Resolve the circular dependency between internal/sbi/processor/processor.go and internal/sbi/nssf.go
 type Nssf interface {
 	Config() *factory.Config
 	Context() *nssf_context.NSSFContext
-	Processor() *Processor
 }
 
 type Processor struct {
