+++ internal/sbi/router.go
@@ -1,4 +1,4 @@
-package server
+package sbi
 
 import (
 	"github.com/gin-gonic/gin"
