+++ pkg/service/init.go
@@ -19,21 +19,23 @@ import (
 	"github.com/free5gc/nssf/internal/logger"
 	"github.com/free5gc/nssf/internal/sbi"
 	"github.com/free5gc/nssf/internal/sbi/consumer"
-	"github.com/free5gc/nssf/internal/sbi/processor"
-	"github.com/free5gc/nssf/internal/sbi/server"
 	"github.com/free5gc/nssf/pkg/factory"
 )
+
+type App interface {
+	Config() *factory.Config
+	Context() *nssf_context.NSSFContext
+}
 
 type NssfApp struct {
 	cfg     *factory.Config
 	nssfCtx *nssf_context.NSSFContext
 
 	wg        sync.WaitGroup
-	sbiServer *server.Server
-	processor *processor.Processor
+	sbiServer *sbi.Server
 }
 
-var _ sbi.Nssf = &NssfApp{}
+var _ App = &NssfApp{}
 
 func NewApp(cfg *factory.Config, tlsKeyLogPath string) (*NssfApp, error) {
 	nssf := &NssfApp{cfg: cfg, wg: sync.WaitGroup{}}
@@ -41,10 +43,7 @@ func NewApp(cfg *factory.Config, tlsKeyLogPath string) (*NssfApp, error) {
 	nssf.SetLogLevel(cfg.GetLogLevel())
 	nssf.SetReportCaller(cfg.GetLogReportCaller())
 
-	processor := processor.NewProcessor(nssf)
-	nssf.processor = processor
-
-	sbiServer := server.NewServer(nssf, tlsKeyLogPath)
+	sbiServer := sbi.NewServer(nssf, tlsKeyLogPath)
 	nssf.sbiServer = sbiServer
 
 	nssf_context.Init()
@@ -58,10 +57,6 @@ func (a *NssfApp) Config() *factory.Config {
 
 func (a *NssfApp) Context() *nssf_context.NSSFContext {
 	return a.nssfCtx
-}
-
-func (a *NssfApp) Processor() *processor.Processor {
-	return a.processor
 }
 
 func (a *NssfApp) SetLogEnable(enable bool) {
